+++ persistence-commons/persistence-commons-api/src/main/java/org/kie/kogito/persistence/api/factory/StorageQualifier.java
