+++ persistence-commons/persistence-commons-api/src/main/java/org/kie/kogito/persistence/api/query/Query.java
